pyproject.toml
[tool.poetry]
name = "library-management"
version = "0.1.0"
description = ""
authors = ["Aarya Gupta <aaryagupta2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = "^0.32.1"
pydantic = "^2.10.3"
python-jose = "^3.3.0"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
psycopg = "^3.2.3"
sqlmodel = "^0.0.22"
psycopg2 = "^2.9.10"
passlib = {extras = ["argon2"], version = "^1.7.4"}
sqladmin = {extras = ["full"], version = "^0.20.1"}
fastapi-admin = "^1.0.4"
aioredis = "^2.0.1"
setuptools = "^75.6.0"
authlib = "^1.3.2"
bcrypt = "^4.2.1"
argon2-cffi = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
